+++ Tests/ZeroMQTests.swift
@@ -28,7 +28,7 @@ class ZeroMQTests: XCTestCase {
         let (major, minor, patch, versionString) = SwiftyZeroMQ.version
         XCTAssertTrue(major == 4, "Major version is 4")
         XCTAssertTrue(minor == 1, "Minor version is 1")
-        XCTAssertTrue(patch == 5, "Patch version is 5")
+        XCTAssertTrue(patch == 6, "Patch version is 6")
         XCTAssertTrue(versionString == "\(major).\(minor).\(patch)")
 
         let frameworkVersion = SwiftyZeroMQ.frameworkVersion